IOS-60 Build w/ Xcode 12 using R2 beta.2
This updates the carthage dependencies build to using xcframeworks wherever possible, therefore this requires carthage 0.37+. Because AudioEngine does not officially support carthage, the workaround we previously had does not work anymore. Therefore this commit pulls NYPLAEToolkit out of carthage entirely, and instead builds it as a git submodule and Xcode subproject for SimplyE/OE.

+++ Simplified/NYPLReaderViewController.m
@@ -482,9 +482,9 @@ spineItemTitle:(NSString *const)title
 {
   if (UIAccessibilityIsVoiceOverRunning()) {
     UIAccessibilityPostNotification(UIAccessibilityPageScrolledNotification,
-                                    [NSString stringWithFormat:NSLocalizedString(@"Page %d of %d", nil),
-                                     pageIndex + 1,
-                                     pageCount]);
+                                    [NSString stringWithFormat:NSLocalizedString(@"Page %lu of %lu", nil),
+                                     (unsigned long)pageIndex + 1,
+                                     (unsigned long)pageCount]);
   }
   
   [self.bottomViewProgressView setProgress:progressWithinBook animated:NO];
